tests/requirements.txt
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
requests>=2.28.0
docker>=6.0.0
pyyaml>=6.0
jinja2>=3.1.0
watchdog>=3.0.0 